Introduce thenAsync method on Future
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57637

I had proposes a similar method in https://github.com/pytorch/pytorch/pull/48790, although that PR was exposing it to Python and thus requires a bit more work. This PR only introduces this method as a C++ API. Python can be added later.

This new method is useful when one wants to use `then` but the callback does perform some async operation itself, and one wants to "reconcile" the future produced inside the callback with the one produced outside.
ghstack-source-id: 129219923

Differential Revision: [D28222809](https://our.internmc.facebook.com/intern/diff/D28222809/)