Allow Future::then to return pre-extracted DataPtrs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58424

In CUDA mode, Future must inspect its value and extract DataPtrs. However some types are not supported, for example the C++/JIT custom classes, which include Message, which is widely used in RPC. Hence for these scenarios we allow the user to perform the custom DataPtr extraction on their own, and pass the pre-extracted DataPtrs.

Note that `markCompleted` already allowed users to pass in pre-extracted DataPtrs, hence this PR simply extends this possibility to the `then` method too.
ghstack-source-id: 129272432

Differential Revision: [D28474880](https://our.internmc.facebook.com/intern/diff/D28474880/)